backend/pyproject.toml
[tool.poetry]
name = "buzzboard api"
version = "0.1.0"
description = "API for BuzzBoard Forum"
authors = ["Ato Toffah <atotoffah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pymongo = "^4.7.2"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

